Using suffix in the namespace to allow for repeating names.
